Web Page Management Using Indexes
Majestic Million CSV is an example of a million record databases.
According to the website, it “a list of the top 1 million website in the
world, ordered by the number of referring subnets.”

As you might imagine, putting this data in a database table, and
performing some database queries on it would be interesting. But as you
also know, we might get a performance hit when using such data. The
solution would be to indexing some of the field. In this project we will
study how to build your website and test how much load it can handle and
improve the performance.